Consolidate all folder paths in package.json
